refactor: separate create folder dialog into new component
