+++ src/renderer/src/components/create-folder-dialog/index.ts
@@ -1,0 +1,1 @@
+export { CreateFolderDialog } from "./components/create-folder-dialog.component";
